// src/server/models/enums/Events.enum.ts
export enum EventType {
  LoginOrCreate = 'reg',
  WinnerUpdate = 'update_winners',
  CreateRoom = 'create_room',
  AddUserToRoom = 'add_user_to_room',
  CreateGame = 'create_game',
  UpdateRoom = 'update_room',
  AddShips = 'add_ships',
  StartGame = 'start_game',
  Attack = 'attack',
  Turn = 'turn',
  Finish = 'finish',
}